requirements.txt
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
django-filter
gunicorn==20.0.4
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
pytest-django==3.9.0